rag-backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG Pipeline - Web Scraping, Embedding Generation & Vector Storage"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "pytest>=8.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.3.5",
    "httpx>=0.25.0",
    "google-generativeai>=0.8.6",
]
